Add a --measure-only option and a --measure-with option
The measure-only option just measures a benchmark and places the output in a
specified temporary file. The file is a means of communication between two
processes. I implemented is using a file but we can use stdout as well instead
of file.

The measure-with option is used to run each benchmark in a separate process. We
specify the path of the benchmark executable itself as an argument to this
option, which is a bit clumsy. The problem is that there is no portable way to
figure the path of the running executable. It is easy to do on Unices but not
on Windows. Therefore the strategy used here is to have the user specify it on
the command line. For example:

stack bench --benchmark-arguments "--measure-with
.stack-work/dist/x86_64-osx/Cabal-2.0.0.2/build/benchmarks/benchmarks
transformation/map/"

With this the measurement is accurate, previous benchmarks do not affect the
later ones and the variance too has got reduced drastically.