Do not discard the first sample during analysis
We anyway use a warmup iteration during measurement. If we want to discard then
we should discard it at the time of measurement itself rather than at the time
of analysis. The analysis process should be offline and we should not have to
remember or guess how the measurement might have been done, it should just
perform analysis on the samples it is given.