Expose minDuration and minSamples to CLI
Allow more control over benchmarking process. This useful in many cases, for
example when we want to use just one iteration or just one sample. Benchmark
results can be wildly different in the first and subsequent iterations because
of a big evaluation happening in the first iteration and then being used by
subsequent ones. We need to provide control to the user rather than just
discarding the results of initial iterations. Otherwise we may be doing unfair
comparisons.