Do not unconditionally discard the first iteration
It is helpful especially in quick-mode to know the results of first iteration.
The first iteration taking too long is indicative of constant applicative forms
(CAFs) or the presence of other pure code that is evaluated only once. It may
or may not be a problem with the way benchmark has been written, quite often it
may be problematic. This option enables the user to investigate and figure out
if that is the case.